Copyedits monitoring topic, further questions in comment
@joel-hamill Here are my proposed edits to this topic. See comments below for more context.

Commit details:

Line 6: This edit may be incorrect. Unclear in source sentence if the additional pluggable stats reporters are using the ``metrics.reporters`` parameter, or if the ``metrics.reporters`` parameter allows the configuration of metrics reporting. I assume it's the latter, and my edit reflects that assumption.  

Line 9-10: Consider adding anchor links in these bullets to the "Global Metrics" and Per-Endpoint Metrics" sections.

Line 11: Including the API endpoint name seems superfluous here. Better to include that in the documentation for that specific metric.

Line 48: This edit may be incorrect. Unclear what "for each one" means in this context. Made an edit assuming that "one" refers to an individual metric. 

Throughout: Line edits to body text to improve readability, remove unnecessary text, tighten language.

Throughout: Recommend putting all metric details, API endpoints, etc. in tables for better use of whitespace and readability. 

I see your other docs don't do this, so I did not make these edits here. Additionally, I wasn't sure what type of table to use--Markdown tables are an abomination, but I'm not sure if HTML will break your stylesheet. Either way, it might be a nice change to include throughout the docs.

Throughout: Recommend altering naming conventions to avoid "master/slave" designations, which can be alienating for some human readers. I would have made this change in my edits, but I wanted to reflect the technical reality of the product until such a change is made throughout. It would be disruptive for the reader to refer to a metrics component as ``kafka.schema.registry:type=master-slave-role`` in code and define what's currently called a slave as a ``worker`` in the docs. 

I recommend changing this language throughout the docs and product.